[CL-2084] Move hooks and services for project folder module to main app
