+++ front/app/containers/Admin/projectFolders/containers/projects/ItemsNotInFolder.tsx
@@ -3,7 +3,7 @@ import { isNilOrError } from 'utils/helperUtils';
 
 // services
 import { PublicationStatus } from 'services/projects';
-import { updateProjectFolderMembership } from '../../../services/projects';
+import { updateProjectFolderMembership } from '../../../../../modules/commercial/project_folders/services/projects';
 
 // hooks
 import useAdminPublications from 'hooks/useAdminPublications';
