--- front/app/modules/commercial/project_folders/services/adminPublications.ts
@@ -1,10 +1,0 @@
-import 'services/adminPublications';
-
-declare module 'services/adminPublications' {
-  export interface IAdminPublicationTypeMap {
-    folder: 'folder';
-  }
-  export interface IQueryParameters {
-    folder?: string;
-  }
-}
