+++ front/app/root.tsx
@@ -14,6 +14,8 @@ import OutletsProvider from 'containers/OutletsProvider';
 import modules from 'modules';
 import history from 'utils/browserHistory';
 
+console.log({ modules });
+
 import {
   createRoutesFromChildren,
   matchRoutes,
